Zyp: Add capability to skip rule evaluation when `disabled: true`
When editing transformation rule files, it is convenient to be able to
quickly disable a certain rule completely. Thus, rules now respect the
new `disabled` attribute on them, which will cause their transformer to
be idempotent, and will also prevent them to be added to the list of
runtime rules.